requirements.txt
Django==2.1
djangorestframework
pillow
psycopg2-binary
markdown
django-filter
scikit-learn
scipy
ipdb
pandas
keras
opencv-python
tqdm
numpy